fix: pinctrl: cp110: fix compilation worning in pinctrl driver
Reviewed-on: http://vgitil04.il.marvell.com:8080/30463
Tested-by: Star_Automation <star@marvell.com>

+++ drivers/pinctrl/mvebu/pinctrl-armada-cp110.c
@@ -649,7 +649,7 @@ static int armada_cp110_pinctrl_probe(struct platform_device *pdev)
 	if (IS_ERR(mpp_base))
 		return PTR_ERR(mpp_base);
 
-	soc->variant = (unsigned) match->data & 0xff;
+	soc->variant = (u64)match->data & 0xff;
 	soc->controls = armada_cp110_mpp_controls;
 	soc->ncontrols = ARRAY_SIZE(armada_cp110_mpp_controls);
 	soc->gpioranges = armada_cp110_mpp_gpio_ranges;
